# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Portfolio Server API",
    "description": "GoLang 기반 게시판 백엔드 서비스",
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Bearer <token>"
    }
  },
  "paths": {
...
    "/articles": {
      "get": {
        "summary": "글 목록 조회",
        "description": "커서 기반 무한 스크롤로 게시글 목록을 조회합니다",
        "tags": ["articles"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "last_id",
            "in": "query",
            "type": "integer",
            "description": "마지막 글 ID"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "default": 20,
            "description": "조회할 개수"
          }
        ],
        "responses": {
          "200": {
            "description": "글 목록 조회 성공"
          },
          "400": {
            "description": "잘못된 요청"
          }
        }
      },
      "post": {
        "summary": "글 작성",
        "description": "새로운 게시글을 작성합니다",
        "tags": ["articles"],
        "security": [{"Bearer": []}],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {"type": "string"},
                "content": {"type": "string"},
                "category_ids": {
                  "type": "array",
                  "items": {"type": "integer"},
                  "description": "카테고리 ID 목록"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "글 작성 성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증 필요"
          }
        }
      }
    },
    "/articles/top/views": {
      "get": {
        "summary": "조회수 TOP 5 글",
        "description": "조회수가 가장 높은 5개의 글을 반환합니다",
        "tags": ["articles"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "성공",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {"type": "integer"},
                  "title": {"type": "string"},
                  "view_count": {"type": "integer"}
                }
              }
            }
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/articles/{id}": {
      "get": {
        "summary": "글 상세 조회",
        "description": "특정 게시글의 상세 정보를 조회합니다",
        "tags": ["articles"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "글 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "글 상세 조회 성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "404": {
            "description": "글을 찾을 수 없음"
          }
        }
      },
      "put": {
        "summary": "글 수정",
        "description": "자신의 게시글을 수정합니다",
        "tags": ["articles"],
        "security": [{"Bearer": []}],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "글 ID"
          },
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {"type": "string"},
                "content": {"type": "string"},
                "category_ids": {
                  "type": "array",
                  "items": {"type": "integer"},
                  "description": "카테고리 ID 목록"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "글 수정 성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증 필요"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "글을 찾을 수 없음"
          }
        }
      },
      "delete": {
        "summary": "글 삭제",
        "description": "자신의 게시글을 삭제합니다",
        "tags": ["articles"],
        "security": [{"Bearer": []}],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "글 ID"
          }
        ],
        "responses": {
          "204": {
            "description": "글 삭제 성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증 필요"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "글을 찾을 수 없음"
          }
        }
      }
    },
...
  },
...
}